refactor: extract calculator engine from state monolith
Decomposes the massive calculator.svelte.ts state into a standalone engine module handling WASM interpolation, error retry, and out-of-range caching. Part of issue #628.
